requirements.txt
openai
faiss-cpu
pandas
numpy
python-dotenv